Create Loading_Mosaicking_SWOT_Raster.py
This is a python script that loads in the SWOT raster data for each cycle id (27-34). The data is then loaded into a geodataframe for all the necessary variables and then mosaicked together with overlapping areas resolved by the median elevation value. The ellipsoid height is then calculated using the provided function.